requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pytest==7.4.4
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
fastapi-cache2==0.2.1
loguru==0.7.2
bcrypt==4.1.2
httpx==0.26.0